backend/requirements.txt
    # absl-py==2.3.1
    # accelerate==1.12.0
    # aiohappyeyeballs==2.6.1
    # aiohttp==3.13.2
    # aiosignal==1.4.0
    # annotated-doc==0.0.4
    # annotated-types==0.7.0
    # anyio==4.11.0
    # astunparse==1.6.3
    # async-timeout==5.0.1
    # attrs==25.4.0
    # backoff==2.2.1
    # bcrypt==5.0.0
    # beautifulsoup4==4.14.2
    # build==1.3.0
    # cachetools==6.2.2
    # certifi==2025.11.12
    # charset-normalizer==3.4.4
    # chromadb==1.3.5
    # click==8.3.1
    # colorama==0.4.6
    # coloredlogs==15.0.1
    # distro==1.9.0
    # durationpy==0.10
    # exceptiongroup==1.3.1
    # fastapi==0.121.3
    # filelock==3.20.0
    # flatbuffers==25.9.23
    # frozenlist==1.8.0
    # fsspec==2025.10.0
    # gast==0.6.0
    # google-auth==2.43.0
    # google-pasta==0.2.0
    # googleapis-common-protos==1.72.0
    # grpcio==1.76.0
    # h11==0.16.0
    # h5py==3.15.1
    # httpcore==1.0.9
    # httptools==0.7.1
    # httpx==0.28.1
    # huggingface-hub==0.36.0
    # humanfriendly==10.0
    # idna==3.11
    # importlib_metadata==8.7.0
    # importlib_resources==6.5.2
    # Jinja2==3.1.6
    # joblib==1.5.2
    # jsonschema==4.25.1
    # jsonschema-specifications==2025.9.1
    # keras==3.12.0
    # kubernetes==34.1.0
    # libclang==18.1.1
    # lxml==6.0.2
    # Markdown==3.10
    # markdown-it-py==4.0.0
    # MarkupSafe==3.0.3
    # mdurl==0.1.2
    # ml_dtypes==0.5.4
    # mmh3==5.2.0
    # mpmath==1.3.0
    # multidict==6.7.0
    # namex==0.1.0
    # networkx==3.4.2
    # numpy==2.2.6
    # oauthlib==3.3.1
    # onnxruntime==1.23.2
    # opentelemetry-api==1.38.0
    # opentelemetry-exporter-otlp-proto-common==1.38.0
    # opentelemetry-exporter-otlp-proto-grpc==1.38.0
    # opentelemetry-proto==1.38.0
    # opentelemetry-sdk==1.38.0
    # opentelemetry-semantic-conventions==0.59b0
    # opt_einsum==3.4.0
    # optree==0.18.0
    # orjson==3.11.4
    # overrides==7.7.0
    # packaging==25.0
    # pandas==2.3.3
    # pillow==12.0.0
    # posthog==5.4.0
    # propcache==0.4.1
    # protobuf==6.33.1
    # psutil==7.1.3
    # pyasn1==0.6.1
    # pyasn1_modules==0.4.2
    # pybase64==1.4.2
    # pydantic==2.12.4
    # pydantic_core==2.41.5
    # Pygments==2.19.2
    # PyMuPDF==1.26.6
    # PyPika==0.48.9
    # pyproject_hooks==1.2.0
    # pyreadline3==3.5.4
    # python-dateutil==2.9.0.post0
    # python-dotenv==1.2.1
    # python-multipart==0.0.20
    # pytz==2025.2
    # PyYAML==6.0.3
    # referencing==0.37.0
    # regex==2025.11.3
    # requests==2.32.5
    # requests-oauthlib==2.0.0
    # rich==14.2.0
    # rpds-py==0.29.0
    # rsa==4.9.1
    # safetensors==0.7.0
    # scikit-learn==1.7.2
    # scipy==1.15.3
    # sentence-transformers==5.1.2
    # shellingham==1.5.4
    # six==1.17.0
    # sniffio==1.3.1
    # soupsieve==2.8
    # starlette==0.50.0
    # sympy==1.14.0
    # tenacity==9.1.2
    # tensorboard==2.20.0
    # tensorboard-data-server==0.7.2
    # tensorflow==2.20.0
    # termcolor==3.2.0
    # threadpoolctl==3.6.0
    # tokenizers==0.22.1
    # tomli==2.3.0
    # torch==2.9.1
    # tqdm==4.67.1
    # transformers==4.57.1
    # typer==0.20.0
    # typing-inspection==0.4.2
    # typing_extensions==4.15.0
    # tzdata==2025.2
    # urllib3==2.3.0
    # uvicorn==0.38.0
    # watchfiles==1.1.1
    # websocket-client==1.9.0
    # websockets==15.0.1
    # Werkzeug==3.1.3
    # wrapt==2.0.1
    # yarl==1.22.0
    # zipp==3.23.0
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
pypdf